move safe-parse to utils
